hub/.requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.62.1
kiwisolver==1.5.0
matplotlib==3.10.9
numpy==2.2.6
packaging==26.2
pillow==12.2.0
pygame-ce==2.5.7
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
